Add possibility to download nodes and edges as CSV
Added a new functionality to download the nodes and edges as two CSV files that can be imported by Gephi. Currently only the code  and document groups (and links between them) will be included in the export.

For more info on the format and import in Gephi, see: https://gephi.org/users/supported-graph-formats/csv-format/